Add new method uninterrupt ability to shut off events
Add the ability to "uninterrupt" the Idler so that interrupting can be
handled manually vs through events.  This allows the creation of a
dialog box which would require some interaction vs simply closing the
dialog box when the user moves the mouse.  Also added some unit tests
around this new functionality